SRP Checkout exercise: move receipt stuff and introduce basket
The Receipt is out of scope for this exercise so I’m moving it out of
the way so it doesn’t distract me.

Missed a commit here and having to batch up my change to introduce the
basket. This is a stepping stone towards not needing to have to have a
variable for each product type.

I’ve introduced a nothing test as I’m nervous that I will break
discounts because I’ve spotted that my discounts use modulus and that
will fire if there are zero items. I might end up giving the customer
money…