SRP Checkout exercise: introduce Catalogue and move findProduct() there
Had to fix a bug in Basket where the same item could not be stored more
than once. Previously this was working as I hadn’t implemented equals()
in my Product class, so the HashSet was relying on identity which
worked because Products were being intialised each time they were
scanned. Since I changed this I needed to fix the bug.